requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
blinker==1.9.0
blis==1.3.0
CacheControl==0.14.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
coverage==7.8.0
cryptography==44.0.3
cycler==0.12.1
cymem==2.0.11
Deprecated==1.2.18
distro==1.9.0
fastapi==0.115.12
filelock==3.18.0
firebase-admin==6.8.0
Flask==3.1.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.5.0
fugashi==1.4.0
future==1.0.0
git-filter-repo==2.47.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
ja_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/ja_core_news_sm-3.8.0/ja_core_news_sm-3.8.0-py3-none-any.whl#sha256=65693016b7cf33ffa845af3523a9c9b8d9aabae85dfc2e46731513a5e03def04
jaconv==0.4.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
line-bot-sdk==3.17.1
loguru==0.7.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
murmurhash==1.0.12
mypy==1.15.0
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
notion-client==2.3.0
numpy==2.2.5
openai==1.77.0
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
preshed==3.0.9
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.5
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
sentry-sdk==2.28.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
SudachiDict-core==20250129
SudachiPy==0.6.10
sympy==1.14.0
tenacity==9.1.2
textblob==0.19.0
thinc==8.3.6
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.1
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
unidic-lite==1.0.8
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
wasabi==1.1.3
watchfiles==1.0.5
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
win32_setctime==1.2.0
wrapt==1.17.2
yarl==1.20.0
